Fix rebase breakage

+++ czip/BitIO.cpp
@@ -173,9 +173,9 @@ int CZip::InputBit(BITFILE *bfile) {
   return (value ? 1 : 0);
 }
 
-ulong CZip::InputBits(BITFILE *bfile, int bitcount) {
-  ulong mask;
-  ulong return_value;
+uint32_t CZip::InputBits(BITFILE *bfile, int bitcount) {
+  uint32_t mask;
+  uint32_t return_value;
 
   mask = 1L << (bitcount - 1);
   return_value = 0;
